backkend/app/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
python-docx==1.1.0
PyPDF2==3.0.1
Pillow==10.2.0
pytesseract==0.3.10
